Move UserLayout to its own file
